enable authorized users make approved requests

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,4 +1,5 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 //date-fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
@@ -6,11 +7,18 @@ import formatDistanceToNow from "date-fns/formatDistanceToNow";
 import { CiTrash } from "react-icons/ci";
 
 const WorkoutDetails = ({ workout }) => {
+  const { user } = useAuthContext();
   const { dispatch } = useWorkoutsContext();
 
   const handleDelete = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
